bug in exact
there is a bug in applying the exact manhattan pdf. Its only being linear atm.

There was also a bug where my zM was a np integer array. casuing any modifications in adjust matrix to to just get rounded. This was fixed(>) by changing the array type.  was probably in the last push though